phase1-pipeline/webserver/requirements.txt
flask==2.3.2
pymysql==1.0.2
mysql-connector-python==8.0.33